scan date is always set to current date
Currently the Fortify parser will set the bom updated date as scan date but if the vulnerablities is zero, then it will set it to java.util.Date. In this scenario, if you do back to back upload, first one will be 6 vulnerabilities and second one with 0 vulnerabilities. We expect it to have 0 vulnerabilities but in this case no. This is because the java.util.Date timezone is different from the Black Duck database timezone. We have fixed this issue by always setting the scan date to current date.